Add tests around reload
