--- tests/integration/mixins/loadable-store/load-all-and-load-record-test.js
@@ -1,81 +1,0 @@
-import { module, test, setupTest } from 'ember-qunit';
-import { waitUntil } from '@ember/test-helpers';
-import { run } from '@ember/runloop';
-import { Model, hasMany, belongsTo } from 'ember-cli-mirage';
-import MirageServer from 'dummy/tests/integration/helpers/mirage-server';
-import LoadableStore from 'ember-data-storefront/mixins/loadable-store';
-
-module('Integration | Mixins | LoadableStore | loadAll and loadRecord', function(hooks) {
-  setupTest(hooks);
-
-  hooks.beforeEach(function() {
-    this.server = new MirageServer({
-      models: {
-        post: Model.extend({
-          comments: hasMany(),
-          tags: hasMany()
-        }),
-        comment: Model.extend({
-          post: belongsTo()
-        }),
-        tag: Model.extend({
-          posts: hasMany()
-        })
-      },
-      baseConfig() {
-        this.resource('posts');
-      }
-    });
-
-    this.store = this.owner.lookup('service:store');
-    this.store.reopen(LoadableStore);
-    this.store.resetCache();
-  }),
-
-  hooks.afterEach(function() {
-    this.server.shutdown();
-  });
-
-  test('loadRecord resolves immediately if its called with no options and the record is already in the store from loadAll, then reloads it in the background', async function(assert) {
-    let serverPost = this.server.create('post', { title: 'My post' });
-    let serverCalls = 0;
-    this.server.pretender.handledRequest = function() {
-      serverCalls++;
-    };
-
-    await run(() => {
-      return this.store.loadAll('post');
-    });
-
-    let post = await run(() => {
-      return this.store.loadRecord('post', serverPost.id);
-    });
-
-    assert.equal(serverCalls, 1);
-    assert.equal(post.get('title'), 'My post');
-
-    await waitUntil(() => serverCalls === 2);
-  });
-
-  test('loadRecord blocks if its called with an includes, even if the record has already been loaded from loadAll', async function(assert) {
-    let serverPost = this.server.create('post', { title: 'My post' });
-    let serverCalls = 0;
-    this.server.pretender.handledRequest = function() {
-      serverCalls++;
-    };
-
-    await run(() => {
-      return this.store.loadAll('post');
-    });
-
-    let post = await run(() => {
-      return this.store.loadRecord('post', serverPost.id, {
-        include: 'comments'
-      });
-    });
-
-    assert.equal(serverCalls, 2);
-    assert.equal(post.get('title'), 'My post');
-  });
-
-});
